Test `ext/mbstring` availability globally.
First, in the constructor, we tested whether the `ext/mbstring`
extension was available or not. This check was made each time the
constructor run. This is useless and we can optimize it. Second, some
static methods were using `ext/mbstring`, so this check was not made and
this situation could lead to unexpected errors.

This patch introduces and testes the `checkMbString` static method. As
its name suggests, it checks whether the `ext/mbstring` extension is
available or not.

We finally use this method globally, after the class declaration, and
throw an exception if the extension is not available.