`ext/iconv` is suggested, no longer required.
First, this patch introduces and testes the `checkIconv` static method
that checks whether the `ext/iconv` extension is available or not.

Second, this method is used in the `Ustring::transcode` static method to
test whether this extension is available. If not, then an exception is
thrown.

Finally, because this is only for one method **and** to allow third
implementation to be used instead of the official `ext/iconv` extension,
this extension is now suggested instead of required in the
`composer.json` file. It removes one (strong) dependency to projects
using `hoa/ustring` if they are sure to not use
`Hoa\Ustring::transcode`.